changed all steps to doubles

+++ KardioKart/OvalRaceTrack.swift
@@ -55,7 +55,7 @@ class OvalRaceTrack: RaceTrack {
     
 
     // MARK: Display
-    override func pointForSteps(steps: Int) -> CGPoint? {
+    override func pointForSteps(steps: Double) -> CGPoint? {
         var percent = min(1, Double(steps) / totalSteps)
         // for an animating track, percent is augmented
         if percent > 1.0 {
